HuggingfaceDeployed/requirements.txt
streamlit
python-dotenv
pydantic
openai
google-generativeai
altair
pandas


